Resolve unaccounted-tax payer names hourly in structures job
Payer names in the structures unaccounted-tax breakdown come from the
eve_name table, which was only populated by the daily job. That job runs
after a large, failure-prone SDE import, so names could lag up to a day
or never resolve, leaving the dropdown showing raw character ids.

Extract the journal-party name resolver into a shared resolveNames module
and run it at the end of the hourly structures job (right after the corp
wallet journals are pulled), so payer names appear within the hour and no
longer depend on the SDE import.
